handled merge conflicts and addressed comments from first PR review
