+++ docs/design/design_with_notes.puml
@@ -58,7 +58,7 @@ package ui{
     class GameController {
       - gameState: GameState
       - display: IGameDisplay
-      - input: IPlayerInput
+      - input: domain.input.IPlayerInput
       - deckFactory: DeckFactory
       - comboValidator: ComboValidator
       + startGame(numPlayers: int): void
@@ -79,7 +79,7 @@ package ui{
 
 package domain{
 
-    interface IPlayerInput {
+    interface domain.input.IPlayerInput {
       + promptCardSelection(player: Player): List<Card>
       + promptNumPlayers(): int
       + promptNope(players: List<Player>): boolean
@@ -112,12 +112,12 @@ package domain{
     class AttackAction implements CardAction
     class ShuffleAction implements CardAction
     class SeeTheFutureAction implements CardAction
-    'nopeAction may not be needed if only prompted via promptNope in IPlayerInput'
+    'nopeAction may not be needed if only prompted via promptNope in domain.input.IPlayerInput'
     class NopeAction implements CardAction
     class DefuseAction implements CardAction
 
     class PlayerInteractionHelper {
-      - input: IPlayerInput
+      - input: domain.input.IPlayerInput
       + stealRandomCard(from: Player, to: Player): void
       + stealNamedCard(from: Player, to: Player, type: CardType): void
     }
@@ -229,7 +229,7 @@ Main ..> GameController
 Main ..> DeckFactory
 
 GameView ..|> IGameDisplay
-GameView ..|> IPlayerInput
+GameView ..|> domain.input.IPlayerInput
 
 Card --> CardType
 Card --> CardAction
@@ -250,12 +250,12 @@ ThreeCatAction --> PlayerInteractionHelper
 
 GameController --> GameState
 GameController --> IGameDisplay
-GameController --> IPlayerInput
+GameController --> domain.input.IPlayerInput
 GameController --> DeckFactory
 GameController --> ComboValidator
 
 ComboValidator --> CardAction
-PlayerInteractionHelper --> IPlayerInput
+PlayerInteractionHelper --> domain.input.IPlayerInput
 
 DeckFactory ..> Deck
 
@@ -273,7 +273,7 @@ note top of IGameDisplay
   to see anyone else's)
 end note
 
-note top of IPlayerInput
+note top of domain.input.IPlayerInput
   Input from the user
   promptCardSelection: asks the player which cards they want to play when it's their turn
   promptNumPlayers: asks at the start how many players are playing
