+++ src/main/java/domain/factory/PlayerInteractionHelper.java
@@ -1,7 +1,7 @@
 package domain.factory;
 
 import domain.enums.CardType;
-import controller.IPlayerInput;
+import domain.input.IPlayerInput;
 import domain.model.Card;
 import domain.model.Player;
