+++ src/main/java/domain/input/IPlayerInput.java
@@ -1,6 +1,7 @@
 package domain.input;
 
 import domain.enums.CardType;
+import domain.enums.PlayerChoice;
 import domain.model.Card;
 import domain.model.Player;
 
@@ -20,4 +21,6 @@ public interface IPlayerInput {
 	CardType promptCardType();
 
 	boolean promptRestart();
+
+	PlayerChoice  promptPlayerChoice();
 }
