+++ src/main/java/ui/GameView.java
@@ -2,6 +2,7 @@ package ui;
 
 import domain.enums.CardType;
 import domain.input.IPlayerInput;
+import domain.enums.PlayerChoice;
 import domain.model.Card;
 import domain.model.GameState;
 import domain.model.Player;
@@ -19,6 +20,9 @@ public class GameView implements IGameDisplay, IPlayerInput {
 	}
 
 	public void showWinner(Player player) {
+	}
+
+	public void showCurrentPlayer(Player player) {
 	}
 
 	public List<Card> promptCardSelection(Player player) {
@@ -48,4 +52,8 @@ public class GameView implements IGameDisplay, IPlayerInput {
 	public boolean promptRestart() {
 		return false;
 	}
+
+	public PlayerChoice promptPlayerChoice() {return null;}
+
+
 }
