+++ src/test/java/domain/factory/DeckFactoryTest.java
@@ -2,7 +2,7 @@ package domain.factory;
 
 import domain.enums.CardName;
 import domain.enums.CardType;
-import controller.IPlayerInput;
+import domain.input.IPlayerInput;
 import domain.model.Card;
 import domain.model.Deck;
 import org.easymock.EasyMock;
